Parsing scenes
