+++ src/resources/parse.ts
@@ -5,6 +5,7 @@ import { parseText } from './parsers/text';
 import { parseEntity } from './parsers/entity';
 import { parseTextureRGBA } from './parsers/texture';
 import { parseBody } from './parsers/body';
+import { parseScene } from './parsers/scene';
 
 const NOP = (resource: Resource) => {
     return resource;
@@ -21,7 +22,7 @@ const ResourceTypes = {
     TXR: { type: 'TXR', description: 'Texture RGBA', parser: parseTextureRGBA },
     M4A: { type: 'M4A', description: 'MPEG 4 Audio', parser: NOP },
     LS1: { type: 'LS1', description: 'LBA1 Scene', parser: NOP },
-    LS2: { type: 'LS2', description: 'LBA2 Scene', parser: NOP },
+    LS2: { type: 'LS2', description: 'LBA2 Scene', parser: parseScene },
     LSP: { type: 'LSP', description: 'LBA Sprite', parser: NOP },
     LSR: { type: 'LSR', description: 'LBA2 Sprite Raw', parser: NOP },
     LBT: { type: 'LBT', description: 'LBA Text Dialog', parser: parseText },
